Moved S3 upload code from OCR service to Upload model.
